add impersonate-browser extra: uses curl_cffi for browser impersonation
This commit tries to more closely mimic web browser behavior in an attempt to
circumvent Cloudflare's bot protection.

If built with the 'impersonate-browser' extra, garminexport will use curl_cffi
and a patched libcurl to produce TLS fingerprints that are identical to those of
a real web browsers.

The prior 'cloudflare' extra is dropped in favor of the new
'impersonate-browser' extra.